pyproject.toml
[project]
name = "2026-personal-expertfinder"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.*"
dependencies = [
    "adapters>=1.2.0",
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.63.0",
    "hdbscan>=0.8.41",
    "jupyter>=1.1.1",
    "llvmlite>=0.43",
    "numba>=0.60",
    "numpy>=1.26,<2.3",
    "pandas>=3.0.0",
    "playwright>=1.58.0",
    "plotly>=6.5.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "umap-learn>=0.5.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
